# openapi.json
{"openapi":"3.1.0","info":{"title":"Stroke Detection Backend","version":"v1.0.0"},"paths":{"/api/v1/users/":{"post":{"tags":["Users"],"summary":"Create User","operationId":"create_user_api_v1_users__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/me":{"get":{"tags":["Users"],"summary":"Get User","operationId":"get_user_api_v1_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"put":{"tags":["Users"],"summary":"Update User","operationId":"update_user_api_v1_users_me_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"tags":["Users"],"summary":"Delete User","operationId":"delete_user_api_v1_users_me_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/assistance/users/":{"get":{"summary":"Get User Timely Assistance","operationId":"get_user_timely_assistance_api_v1_assistance_users__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/auth/login":{"post":{"tags":["Auth"],"summary":"Login","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analyze_balance":{"post":{"tags":["Detection"],"summary":"Analyze Balance Endpoint","operationId":"analyze_balance_endpoint_api_v1_analyze_balance_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SensorData"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/analyze_speech":{"post":{"tags":["Detection"],"summary":"Analyze Speech Endpoint","operationId":"analyze_speech_endpoint_api_v1_analyze_speech_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_analyze_speech_endpoint_api_v1_analyze_speech_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/patients/":{"post":{"tags":["Patients"],"summary":"Create Patient Profile","operationId":"create_patient_profile_api_v1_patients__post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_patient_profile_api_v1_patients__post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/patients/me":{"get":{"tags":["Patients"],"summary":"Get Patient Profile","operationId":"get_patient_profile_api_v1_patients_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientModel"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/patients/history/me":{"get":{"tags":["Patients"],"summary":"Get Detection History","operationId":"get_detection_history_api_v1_patients_history_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Detection"},"type":"array","title":"Response Get Detection History Api V1 Patients History Me Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/ping":{"get":{"summary":"Root","operationId":"root_ping_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"BMI":{"properties":{"height_cm":{"type":"number","title":"Height Cm"},"weight_kg":{"type":"number","title":"Weight Kg"}},"type":"object","required":["height_cm","weight_kg"],"title":"BMI"},"Body_analyze_speech_endpoint_api_v1_analyze_speech_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_analyze_speech_endpoint_api_v1_analyze_speech_post"},"Body_create_patient_profile_api_v1_patients__post":{"properties":{"photo":{"type":"string","format":"binary","title":"Photo"},"voice_sample":{"type":"string","format":"binary","title":"Voice Sample"},"height_cm":{"type":"number","title":"Height Cm"},"weight_kg":{"type":"number","title":"Weight Kg"}},"type":"object","required":["photo","voice_sample","height_cm","weight_kg"],"title":"Body_create_patient_profile_api_v1_patients__post"},"Detection":{"properties":{"_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"user_id":{"type":"string","title":"User Id"},"detected_at":{"type":"string","format":"date-time","title":"Detected At"},"model_version":{"type":"string","title":"Model Version","default":"v1.0"},"input_type":{"type":"string","enum":["balance","slurred_speech","eye","comprehensive"],"title":"Input Type"},"balance_test":{"anyOf":[{"$ref":"#/components/schemas/DetectionResult"},{"type":"null"}]},"slurred_speech_test":{"anyOf":[{"$ref":"#/components/schemas/DetectionResult"},{"type":"null"}]},"eye_test":{"anyOf":[{"$ref":"#/components/schemas/DetectionResult"},{"type":"null"}]},"overall_result":{"anyOf":[{"type":"string","enum":["stroke_detected","normal"]},{"type":"null"}],"title":"Overall Result"},"additional_notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Additional Notes"}},"type":"object","required":["_id","user_id","input_type","balance_test","slurred_speech_test","eye_test","overall_result"],"title":"Detection"},"DetectionResult":{"properties":{"confidence_score":{"type":"number","title":"Confidence Score"},"result":{"type":"string","enum":["stroke_detected","normal"],"title":"Result"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","required":["confidence_score","result"],"title":"DetectionResult"},"EmergencyContact":{"properties":{"name":{"type":"string","title":"Name"},"relation":{"type":"string","title":"Relation"},"phone":{"type":"string","title":"Phone"}},"type":"object","required":["name","relation","phone"],"title":"EmergencyContact"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MedicalHistoryEntry":{"properties":{"condition":{"type":"string","title":"Condition"},"diagnosed_at":{"type":"string","format":"date-time","title":"Diagnosed At"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes","default":""}},"type":"object","required":["condition","diagnosed_at"],"title":"MedicalHistoryEntry"},"PatientModel":{"properties":{"_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"user_id":{"type":"string","title":"User Id"},"photo":{"type":"string","format":"binary","title":"Photo"},"voice_sample":{"type":"string","format":"binary","title":"Voice Sample"},"bmi":{"$ref":"#/components/schemas/BMI"},"medical_history":{"items":{"$ref":"#/components/schemas/MedicalHistoryEntry"},"type":"array","title":"Medical History","default":[]}},"type":"object","required":["_id","user_id","photo","voice_sample","bmi"],"title":"PatientModel"},"SensorData":{"properties":{"accel":{"items":{"items":{"type":"number"},"type":"array"},"type":"array","title":"Accel"},"gyro":{"items":{"items":{"type":"number"},"type":"array"},"type":"array","title":"Gyro"}},"type":"object","required":["accel","gyro"],"title":"SensorData"},"UserCreate":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"role":{"type":"string","enum":["patient","admin"],"title":"Role","default":"patient"},"emergency_contacts":{"items":{"$ref":"#/components/schemas/EmergencyContact"},"type":"array","title":"Emergency Contacts","default":[]}},"type":"object","required":["name","email","password"],"title":"UserCreate"},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserLogin"},"UserOut":{"properties":{"_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"role":{"type":"string","title":"Role"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"emergency_contacts":{"items":{"$ref":"#/components/schemas/EmergencyContact"},"type":"array","title":"Emergency Contacts","default":[]}},"type":"object","required":["_id","name","email","role","created_at"],"title":"UserOut"},"UserUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"role":{"anyOf":[{"type":"string","enum":["patient","admin"]},{"type":"null"}],"title":"Role"},"emergency_contacts":{"anyOf":[{"items":{"$ref":"#/components/schemas/EmergencyContact"},"type":"array"},{"type":"null"}],"title":"Emergency Contacts"}},"type":"object","title":"UserUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/auth/login"}}}}}}